+++ Cargo.toml
@@ -74,13 +74,13 @@ unexpected_cfgs = { level = "warn", check-cfg = [
 # Dependencies that should be kept in sync through the whole workspace
 [workspace.dependencies]
 k256 = { version = "0.14.0-pre.9", features = ["arithmetic", "critical-section", "precomputed-tables", "serde", "ecdsa", "hash2curve", "alloc"], default-features = false }
-mpc = { git = "https://github.com/dwallet-labs/inkrypto", rev = "29d2bcf"}
-proof = { git = "https://github.com/dwallet-labs/inkrypto", rev = "29d2bcf"}
-class_groups = { git = "https://github.com/dwallet-labs/inkrypto", rev = "29d2bcf", features = ["threshold"] }
-commitment = { git = "https://github.com/dwallet-labs/inkrypto", rev = "29d2bcf" }
-twopc_mpc = { git = "https://github.com/dwallet-labs/inkrypto", features = ["secp256k1", "class_groups"], rev = "29d2bcf"}
-group = { git = "https://github.com/dwallet-labs/inkrypto", features = ["os_rng"], rev = "29d2bcf"}
-homomorphic_encryption = { git = "https://github.com/dwallet-labs/inkrypto", rev = "29d2bcf"}
+mpc = { git = "https://github.com/dwallet-labs/cryptography-private", rev = "bc0464c"}
+proof = { git = "https://github.com/dwallet-labs/cryptography-private", rev = "bc0464c"}
+class_groups = { git = "https://github.com/dwallet-labs/cryptography-private", rev = "bc0464c", features = ["threshold"] }
+commitment = { git = "https://github.com/dwallet-labs/cryptography-private", rev = "bc0464c" }
+twopc_mpc = { git = "https://github.com/dwallet-labs/cryptography-private", features = ["secp256k1", "class_groups"], rev = "bc0464c"}
+group = { git = "https://github.com/dwallet-labs/cryptography-private", features = ["os_rng"], rev = "bc0464c"}
+homomorphic_encryption = { git = "https://github.com/dwallet-labs/cryptography-private", rev = "bc0464c"}
 anyhow = "1.0.71"
 arc-swap = { version = "1.5.1", features = ["serde"] }
 assert_cmd = "2.0.6"
