+++ crates/dwallet-mpc-centralized-party/Cargo.toml
@@ -6,7 +6,7 @@ version.workspace = true
 [dependencies]
 mpc.workspace = true
 twopc_mpc.workspace = true
-commitment = { git = "https://github.com/dwallet-labs/inkrypto", rev = "29d2bcf" }
+commitment = { git = "https://github.com/dwallet-labs/cryptography-private", rev = "bc0464c" }
 message-digest = { path = "../message-digest" }
 class_groups.workspace = true
 group.workspace = true
